Rename file for consistency.
